refactor: rename _build_mixin.py → _build.py
Final rename in the builder package reorganization. All mixin files
now have clean names: _grammar.py, _build.py, _output.py.
